# swagger.json
{
  "info": {
    "title": "CRUD API",
    "version": "1.0.0",
    "description": "A sample API"
  },
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      },
      "XApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Lista os usuários",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuários"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Criar um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "lucas@teste.com.br"
                  },
                  "password": {
                    "type": "string",
                    "example": "teste123"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "XApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna usuário criado"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retorna User",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "security": [
          {
            "XApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON de User"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Atualiza um usuário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "lucas@teste.com.br"
                  },
                  "password": {
                    "type": "string",
                    "example": "teste123"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "XApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna usuário atualizado"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Deleta um usuário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "XApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna o usuário"
          },
          "404": {
            "description": "usuário não encontrada"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "User",
      "description": "API de usuários."
    }
  ]
}